(svn r22837) [1.1] -Backport from trunk: - Fix: [NewGRF] Always draw NewGRF supplied texts with a default colour (r22725) - Fix: [NewGRF] Do not restrict AdvVarAct2 to 255 operations (r22723) - Fix: If there is no point in opening the rail/air toolbar, do not open it for people who use hotkeys either rather than only for those using GUI elements (r22716, r22715, r22714) - Fix: [NoAI] Allow AIAirport::GetNoiseLevelIncrease() also for expired airports [FS#4704] (r22710)

+++ src/newgrf_spritegroup.h
@@ -197,7 +197,7 @@ struct DeterministicSpriteGroup : SpriteGroup {
 
 	VarSpriteGroupScope var_scope;
 	DeterministicSpriteGroupSize size;
-	byte num_adjusts;
+	uint num_adjusts;
 	byte num_ranges;
 	DeterministicSpriteGroupAdjust *adjusts;
 	DeterministicSpriteGroupRange *ranges; // Dynamically allocated
